+++ packages/ui-components/web/src/index.ts
@@ -14,7 +14,7 @@ export { default as Panel } from './Components/Panel';
 // export { default as SelectionList } from './Components/SelectionList';
 // export { default as SettingsPanelFooter } from './Components/SettingsPanelFooter';
 // export { default as SettingsSection } from './Components/SettingsSection';
-// export { default as Slider } from './Components/Slider';
+export { default as Slider } from './Components/Slider';
 // export { default as SliderWithInput } from './Components/SliderWithInput';
 // export { Tab, Tabs } from './Components/Tabs';
 // export { default as TextInput } from './Components/TextInput';